refactor(ios): relocate simctl provisioning util out of maestro-ios-driver
Move the four iOS-simulator provisioning types DeviceService depends on
(LocalSimulatorUtils, LocalIOSDevice, SimctlList, DeviceCtlResponse) plus
their in-module closure (CommandLineUtils, Mp4DurationNormalizer) from
maestro-ios-driver's util.* package into maestro-client's new
maestro.device.ios package. Provisioning (sim boot/discovery) is legitimately
kept; only iOS xctest execution is the legacy being removed.

Drops the dead launchUITestRunner method (xctest execution, referenced the
deleted xctestLogFile helper) and the unused IOSDeviceType enum. Relocates the
two tests that cover this code and their mp4 fixture.
