+++ maestro-client/src/main/resources/screenrecord.sh
