+++ maestro-client/src/test/resources/mp4/short_media.mov
